explore: stop truncating the catalogue at 60 swims
The page never sent p_limit, so search_event_editions applied its own
default of 60. With 185 events live that looked like the crawler had
stopped finding things. Now asks for the RPC's ceiling of 200, and when
a result set hits that exactly it says "showing the first 200" rather
than reporting 200 as the total - a truncated list that looks complete
is worse than one that admits it.

Also fixes the requeue migration's make_interval cast (row_number()
returns bigint).

+++ sql/2026-08-04_requeue-sources-after-worker-fix.sql
@@ -71,7 +71,7 @@ BEGIN;
 -- the organisers' servers too — 36 simultaneous crawls is a spike they
 -- have no reason to absorb.
 UPDATE discovery_sources s
-   SET next_run_at = now() + (make_interval(mins => (r.rn - 1) * 2)),
+   SET next_run_at = now() + (make_interval(mins => (((r.rn - 1) * 2))::int)),
        updated_at  = now()
   FROM (
     SELECT id, row_number() OVER (
